Use catalina.sh start verb instead of run
Handling a graceful stop of the container requires trapping and 
forwarding the stop signal to catalina.sh which doesn't work well when 
it's running as a job. We trap the signal and use catalina.sh to stop 
the application instead.

catalina.sh does not properly write its PID file, so it's not checked. 
It should be, but I can't get it to work.